DPL: support data-less timers
So far timers still needed some input data to be triggered, because only
data was able to create an entry in the TimesliceIndex. This expands the
record expiration logic so that it can also create new timeslices, if
needed, hence adding the support for standalone timers.

Notice this could have also been possible via the idle callback
mechanism, but we prefer to avoid it so that the idle callback is
available to the user even if timers are needed.