Implementation of methods to store TObjects in TFiles in the CCDB. (#2022) (QC-127)
There are new methods to store and retrieve TObjects to and from a TFile in the CCDB. I added the tests for it. It should have no impact on the existing code as it is just extra methods.

CAVEAT : until ROOT 6.18 the minimum size of the TFiles will be 2MB. It will automatically switch to blocks of 1KB when we upgrade ROOT.