Introducing configurable dispatch policy for devices
`DispatchPolicy` describes on per-device basis when completed objects are dispatched
for sending, either at the end of computation - the default and current behavior - or
immediately when ready. Depending on the `DispatchPolicy` operation, the `DataProcessingDevice`
configures the `MessageContext` with a dispatch callback. The dispatch policy can be customized
in the usual DPL manner.