DPL Analysis: improve performance of ColumnIterator
ColumnIterator is now responsible only of switching from
one chunk to the other. The actual iteration index, for
all the columns, is now hold by the RowView itself. We
also check for boundary conditions in the chunk in the accessor,
rather than when doing the increment, since the check is
mostly folded in the subsequent dereference by the branch
prediction logic.

Also we introduce a policy to distinguish the two cases in which we
are iterating over chunked columns (or not). The latter being much
faster.