Updating the definition of RAWDataHeader
The references for RDH v4 and v5 are now in the wp6-doc repo
https://gitlab.cern.ch/AliceO2Group/wp6-doc/blob/master/rdh/RDHV4.md
https://gitlab.cern.ch/AliceO2Group/wp6-doc/blob/master/rdh/RDHV5.md

The previous changes to the RAWDataHeader v3 definition have committed
the update to v4 already, simply renaming the struct. Version 3 and 4
are almost identical, v3 is only missing a few fields which are filled
by the CRU, we keep v3 as a simple alias to v4.

Adding the v5 definition, which contains only one trigger information
set, the length of a few filds has been adjusted, for details see
https://indico.cern.ch/event/818188/#26-cru
https://indico.cern.ch/event/791295/#1-wp6-rdh-v5

Version 5 is not yet implemented in hardware, keeping v4 as default for now.