Split detector and geometry constant parameters (#2475)
* Split detector and geometry constant parameters

Some detector constants are used everywhere, while some others are used only when geometrical information is needed.
Splitting the two series of constants allows for a more efficient handling of the two use case.
This also allows to move some general information outside of Geometry.h, so that one does not need to depend from Simulation just to have those values.
Finally, in the current implementation the functions are grouped inside namespaces. This is better than implementing them as static methods of one class, as was done before.