DPL I/O API: adding support for spectator vector
Changing API for InputRecord::get<std::vector<T>>
Return type changed to vector<T const, o2::pmr::SpectatorAllocator<T const>>

A spectator vector for unserialized messages of messageable type is created
using the raw input memory without copy.
Depending on the serialization method of the incoming message, the spectator
vector is created with raw memory of the message or the deserialized object
as underlying resource

A unique instance of the spectator memory resource is kept in the
InputRecord and valid for the lifetime of the record.