DPL workflow creation: pass original argc,argv as part of ConfigContext (#3222)
So far, during topology construction, we could not know
if the command line was invoked merely to list program options.
This is problematic since topology construction might need to
do some computation / read files / etc. which influence the workflow.
In case of --help this should not be needed.

This commit tries to address this issue by adding additional information
to the ConfigContext which is passed to the workflow construction function.

The principle is applied to the digitizer workflow, which for the first
time prints out the expected output when invoked with
`o2-sim-digitizer-workflow --help` (up until now one needed a GRP file
for this to work).