Arranging data with respect to sector number instead of subspec
Sort DataRef objects for the different sectors by sector number. This will
become necessary if multiple sectors are sent over the same route in a
single timeslice.

Also defining helper methods in the header file as 'static inline' to avoid
conflicts if the header file is inlcuded in both an executable and a library.