Temporary change of raw utils SubSpec definition
given that some detectors (e.g. TOF, TPC) at the moment cannot guarantee a unique
(within a given run) mapping FEEID <-> {cruID, linkID, EndPoint}, the
defined by the DataDistribution may mix data of different links (FEEDs) to single DPL input.

To avoid this, the  definition is temporarily changed to a 32 bit hash
code (Fletcher32). When the new RDHv6 (with source ID) will be put in production, both
DataDistribution and RawFileReader will be able to assign the  in a
consistent way, as a combination of FEEID and SourceID.