Sanity check for O2 headers on the header stack
Checking if the header version matches the implementation.
There is no support for multiple versions of a particular header at
the moment. The minimum is to require the version to match the current
implementation. Header size can not be used for sanity check because the
data model includes support for variable-size headers.

Also checking consistency of the header stack when advancing to the next
header by checking that it starts with the BaseHeader information at the
offset given by the size of the current header.