DPL: fix forwarding to make sure messages arrive together (#3627)
Messages on the same channel were supposed to be collected when forwarding
to next processor. This is to make sure that messages receive together are
also forwarded together. The optimization to use the pointer to channel name
string as key in the map could not work because every forward entry has it's
own string. As a consequence, all messages have been sent separately.