DPL: proper Services API (#3889)
This finally implements Services as originally conceived:

* Each dataprocessor has a list of required services specification.
  The specification will allow creating the service and (optionally)
  configuring it, attaching them to the provided Context.
* Required services are instanciated before running, without having to
  hardcode the list.

Eventually this will be extended to support transitions in a multithreaded
environment (e.g. some services might have to be created on a per thread
basis).

By default the old list of hardcoded services is provided. Users can override
that via the usual customise mecanism.