DPL: progress towards multithreading support (#3983)
The actual data processing is now somewhat reentrant, as it's performed
by static functions which act on a DataProcessingContext object.
This will allow in the future to have different contexes for different threads
making sure that each context either points to a thread safe component,
or to a per thread copy.