DPL: make ServiceRegistry thread safe and aware
the ServiceRegistry has now knowledge of the fact that the services
inside have different behavior depending on their thread safety. In
particular there are three kind of services:

* ServiceKind::Serial unique instance, which will have to be locked.
* ServiceKind::Global unique instance, renetrant and thread safe.
* ServiceKind::Stream multiple instances, one per thread.

For the time being this only translates in returning the same pointer for
Serial / Global services. In a near future the function will return a smart
pointer which will provide functionalities for locking / unlocking access to
the service itself.