[rANS] automatic calculation of rescaling factor
the rescaling factor determines the precision of the rANS compression
algorithm SymbolTable and this it's compression efficiency and
performance.
Until now it had to be set manually. The parameter now has become
optional and the optimal precision is automatically estimated based on
the FrequencyTable that is used to compress the data.