Add abs.charge data member to TrackPar
This is needed for uniform treatment of tracks with non-standard charge: 0 (for V0s) and e.g. 2 for hypernuclei.
In the aliroot AliExternalTrackParam this was treated by derived classes using virtual methods, which we don't use in O2.
The meaning of mP[kQ2Pt] remains exactly the same, except for q=0 case: in this case the mP[kQ2Pt] is just an alias to
1/pT, regardless of its sign, and the getCurvature() will 0 (because mAbsCharge is 0).
The methods returning lab momentum or its combination account for eventual q>1.