DPL: avoid busy looping when in Idle state (#4481)
Under certain conditions (most notably when an enumeration was provided and
the device switched to "Idle" state) the libuv integration mistaken ly
considered as activity any further enumeration, resulting in a busy loop. This
makes sure we always UV_ONCE (i.e. wait for an external event) when in
StreamingState::Idle.