CCDB API: generalization of content retrival (part 1)
This commit achieves the following:

a) support for cases in which the CCDB server does not serve
   content directly but redirects to other resources

b) support for objects on ALIEN resource

Changes are offered only for retrieval of objects via the strongly
typed interfaces (no new features for interfaces based on TObject*).

A new test is added - querying from a server with ALIEN backend.

This concerns: https://alice.its.cern.ch/jira/browse/O2-1809

The PR is potentially incomplete. Further changes in other places
or tools likely.