Adding classes to calculate the space charge distortions and corrections (#4618)
* Adding classes to calculate the space charge distortions and corrections

The distortions and corrections can be calculated by using the potential
from a 3D histogram or an analytical formula as an input. These classes are
much faster and consistent than the old methods located in GPU/TPCSpaceChargeBase.
The old folder (GPU/TPCSpaceChargeBase) where all the previous classes were stored are deleted
and all other classes and macros which used the old spacecharge classes
are modified to be able to use the new spacecharge classes.

* moved CGAL includes to base class

* moved global objects in NearestNeighbour src file to own class

* adding more grid sizes for calculation of distortions/corrections

* fixed createResidualDistortionObject.C macro for new space charge classes and moved to space charge folder

* clang format

* changed macro calculateDistortionsCorrections.C to compile only

* fixed z coordinate in distortElectron

* removed CGAL package and replaced nearest neighbour searching by approximation of nearest neighbour