Avoid data copy in the DPL output proxy (#6353)
This is a demonstrator to use forward routes of the DataProcessingDevice,
which allows to remove handling of messages in the processing callback
There is no need to do processing of the messages except keeping a copy
of the last DataProcessingHeader which is needed to forward EOS.

The remaining issue how to define these external forward routes is being
discussed. ServiceRegistry is providing access to DeviceSpec, but this is a
const object by design. Once the appropriate API is ready, we can avoid
the tenporary const_cast.

test_ExternalFairMQDeviceWorkflow is now using specifyFairMQDeviceMultiOutputProxy
to better reflact the use cases, but these methods need to be consolidated.