CcdbApi: Option to build/use local cache on the fly
Offering a mode - useful for GRID - in which
we save downloaded objects into a given CCDB disc
store, thereby creating a local cache as we go.

This local store can be used to reduce the number
of network queries, as well as to know exactly which
objects have been requested by whom.

The cache can also be distributed with the software stack
in case someone doesn't have network access or credentials.

The usage is:

```
export ALICEO2_CCDB_LOCALCACHE=$pwd/.ccdb
process1
process2
```

where `process1` and `process2` would both build and use
the local cache located in the .ccdb folder (depending on the order).

Limitations:
- only to be used when all processes query for the same
  timestamp (typically the case for a single GRID job)